PWA-2291: Renamed redux `hook` helper to `mutable`
